fix formating and create transformations_common
