+++ tests/tests_integration/test_api/test_transformation_jobs.py
@@ -97,6 +97,7 @@ class TestTransformationJobsAPI:
             and job.error is None
             and job.ignore_null_fields
         )
+        await asyncio.sleep(0.5)
         retrieved_transformation = COGNITE_CLIENT.transformations.retrieve(id=new_transformation.id)
 
         assert retrieved_transformation.running_job is not None and retrieved_transformation.running_job.id == job.id
